Refactor: Grade 카프카 이벤트 관련 DTO 을 common 모듈로 이동
