+++ GlowGrow-common/src/main/java/com/tk/gg/common/kafka/grade/GradeForReservationEventDto.java
@@ -1,4 +1,4 @@
-package com.tk.gg.reservation.infrastructure.messaging;
+package com.tk.gg.common.kafka.grade;
 
 
 import lombok.Builder;
